refactor(capabilities): modularize seam crates
Split AI, governance, and bot-channel seam crates into responsibility-focused internal modules while preserving their top-level public API and tested behavior. This keeps capability ports local and readable without adding new workspace/package boundaries.
